Add dashboard to menu
